% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainColoring.R
\name{domain_coloring}
\alias{domain_coloring}
\title{Generate domain colorings of a given function of a complex variable}
\usage{
domain_coloring(
  f,
  domain = 1.5,
  n = 200,
  a = 0.01,
  filename = NULL,
  saveJPEG = FALSE
)
}
\arguments{
\item{f}{single-variable pre-defined function of complex variable z}

\item{domain}{the symmetric interval half-radius for both x and y grids}

\item{n}{number of points to use in grids}

\item{a}{luminosity}

\item{filename}{filename to save}

\item{saveJPEG}{boolean whether to save plot or not}
}
\value{
ggplot plot
}
\description{
{A basic function to compute domain colorings of functions of a single complex variable}
}
\details{
{Uses HSV conversion scheme to color a complex mapping based off its argument and modulus.}
}
